+++ src/Relations/HasMany.php
@@ -16,7 +16,6 @@ namespace Blast\Orm\Relations;
 
 use Blast\Orm\Entity\EntityAdapterLoaderTrait;
 use Blast\Orm\Entity\EntityHydratorInterface;
-use Blast\Orm\Entity\Definition;
 use Blast\Orm\Query;
 
 class HasMany implements RelationInterface
